Generator Try/Catch and Tk Crash Fix
Added a try/catch statement to the generator.py file when the
Generator object tries to open the calibration files. This
should help catch the case where the file does not exist or we
are looking in the wrong folder.

Also implemented a fix for the GUI crashing. Instead of using a
second thread to check the queue for an incoming message, I now
use the root.after() command recursively to check the queue every
half a second. Will try this later today to see if it works.